4819 PublishModalLayout
